# openapi.yaml
openapi: 3.0.3
info:
  title: Zapiki - Zero-Knowledge Proof as a Service API
  description: |
    Zapiki is a universal Zero-Knowledge Proof platform that abstracts cryptographic complexity
    behind a simple REST API. Generate and verify cryptographic proofs without deep knowledge
    of cryptography.

    **Core Concept:** "Stripe for Zero-Knowledge Proofs"

    ## Features
    - Multiple proof systems (Commitment, Groth16, PLONK, STARK)
    - Pre-built templates for common use cases
    - Synchronous & asynchronous proof generation
    - Batch proof generation (up to 100 proofs per request)
    - Prometheus metrics for monitoring and observability
    - Simple REST API with API key authentication

    ## Proof Systems

    ### Commitment (Fast - ~50ms)
    Simple hash-based commitments with digital signatures. Perfect for timestamping and
    simple proofs of knowledge.

    ### Groth16 (zk-SNARK - ~30s)
    Zero-knowledge Succinct Non-interactive ARgument of Knowledge with trusted setup.
    Small proof size, fast verification.

    ### PLONK (zk-SNARK - ~35s)
    Universal setup zk-SNARK. No per-circuit trusted setup required. More flexible than Groth16.

    ### STARK (Transparent - ~40s)
    Scalable Transparent ARgument of Knowledge. No trusted setup required, quantum-resistant,
    hash-based proofs. Larger proof size but completely transparent and verifiable by anyone.

    ## Authentication
...

    ## Rate Limiting
    - Free tier: 100 requests/minute
...

...
  contact:
...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/portal/overview:
    get:
      tags:
        - Portal
      summary: Customer usage/failure/cost overview
      description: Consolidated pilot dashboard data for authenticated users.
      responses:
        '200':
          description: Portal overview data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalOverview'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Get your API key from the Zapiki dashboard.

        Example: `X-API-Key: test_zapiki_key_1230ab3c044056686e2552fb5a2648cd`
  schemas:
...
    AuditEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        event_type:
          type: string
        resource:
          type: string
        resource_id:
          type: string
        request_id:
          type: string
        success:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
...
    PortalUsageRow:
      type: object
      properties:
        proof_system:
          type: string
        operation:
          type: string
        count:
          type: integer
        successes:
          type: integer
        failures:
          type: integer
        estimated_cost_usd:
          type: number
          format: double
    PortalOverview:
      type: object
      properties:
        since:
          type: string
          format: date-time
        lookback_days:
          type: integer
        kpis:
          type: object
          properties:
            total_operations:
              type: integer
            total_failures:
              type: integer
            failure_rate:
              type: number
              format: double
            estimated_cost_usd:
              type: number
              format: double
        usage_rows:
          type: array
          items:
            $ref: '#/components/schemas/PortalUsageRow'
        plan_limits:
          type: object
          additionalProperties:
            type: integer
        recent_failed_events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        cost_model:
          type: object
          additionalProperties: true
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error details
  responses:
    UnauthorizedError:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: 401
...